// platform/frontend/src/features/operations/types.ts
export interface Lead {
  id: string;
  firstName: string;
  lastName: string;
  email: string;
  mobile: string;
  panNumber: string;
  loanType: string;
  loanAmount: number;
  assignedTo: string;
  status: string;
  customerId: string;
  createdAt: string;
  profession?: string;
  netMonthlySalary?: number;
  gender?: string;
  maritalStatus?: string;
  dob?: string;
  alternateContact?: string;
  whatsappNo?: string;
  officialEmail?: string;
  currentAddressLine1?: string;
  currentAddressLine2?: string;
  currentState?: string;
  currentDistrict?: string;
  currentCity?: string;
  currentPincode?: string;
  residenceType?: string;
  permanentAddressLine1?: string;
  permanentAddressLine2?: string;
  permanentState?: string;
  permanentDistrict?: string;
  permanentCity?: string;
  permanentPincode?: string;
  jobType?: string;
  designation?: string;
  modeOfSalary?: string;
  officeAddress?: string;
  officeState?: string;
  officeDistrict?: string;
  officeCity?: string;
  officePincode?: string;
  existingEmi?: number;
  hasPriorPersonalLoan?: boolean;
  opsNotes?: string;
}